Add cleaned customer_segmentation module
