+++ customer_segmentation/requirements.txt
@@ -1,0 +1,8 @@
+pandas==2.2.3
+numpy==1.23.1
+scikit-learn==1.2.2
+matplotlib==3.10.1
+matplotlib-inline==0.1.6
+seaborn==0.13.2
+python-dateutil==2.9.0.post0
+scipy==1.9.0
